+++ src/Repository/ObjectHelper.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace srag\Plugins\SrLifeCycleManager\Repository;
 
@@ -19,39 +21,9 @@ trait ObjectHelper
     protected $tree;
 
     /**
-     * @inheritDoc
-     */
-    public function getRepositoryObjects(int $ref_id = 1) : Generator
-    {
-        $routine_candidates_and_container = $this->tree->getChildsByTypeFilter(
-            $ref_id,
-            ['crs', 'cat', 'grp', 'fold', 'itgr', 'svy']
-        );
-        
-        if (empty($routine_candidates_and_container)) {
-            return;
-        }
-        
-        foreach ($routine_candidates_and_container as $candidate_or_container) {
-            $candidate_or_container_ref_id = (int) $candidate_or_container['ref_id'];
-            if (in_array($candidate_or_container['type'], IRoutine::ROUTINE_TYPES, true)) {
-                try {
-                    yield $candidate_or_container_ref_id => ilObjectFactory::getInstanceByRefId(
-                        $candidate_or_container_ref_id
-                    );
-                } catch (Exception $exception) {
-                    continue;
-                }
-            } else {
-                yield from $this->getRepositoryObjects($candidate_or_container_ref_id);
-            }
-        }
-    }
-
-    /**
      * @inheritdoc
      */
-    public function getParentId(int $ref_id) : int
+    public function getParentId(int $ref_id): int
     {
         // type-cast is necessary, the phpdoc comment is wrong.
         return (int) $this->tree->getParentId($ref_id);
@@ -64,7 +36,7 @@ trait ObjectHelper
      * @param int $ref_id
      * @return array|null
      */
-    protected function getParentIdsRecursively(int $ref_id) : array
+    protected function getParentIdsRecursively(int $ref_id): array
     {
         if (1 === $ref_id) {
             return [];
@@ -104,8 +76,19 @@ trait ObjectHelper
      * @param int $ref_id
      * @return string
      */
-    protected function getParentIdForSqlComparison(int $ref_id) : string
+    protected function getParentIdForSqlComparison(int $ref_id): string
     {
         return (string) ($this->tree->getParentId($ref_id) ?? 0);
+    }
+
+    /**
+     * Returns all repository object types which can hold further
+     * objects (children).
+     *
+     * @return string[]
+     */
+    protected function getContainerObjectTypes(): array
+    {
+        return ['crs', 'cat', 'grp', 'fold', 'itgr'];
     }
 }